+++ CleverPush/Source/NSDictionary+SafeExpectations.h
@@ -28,7 +28,7 @@
  @param numberFormatter The formatter to use to parse the number if the object found on the key is a string
  @returns the resulting number. If the result is not a NSNumber and can't converted to one, it returns nil
  */
-- (NSNumber *)numberForKey:(id)key usingFormatter:(NSNumberFormatter *)numberFormatter;
+- (NSNumber *)cleverPushNumberForKey:(id)key usingFormatter:(NSNumberFormatter *)numberFormatter;
 
 /**
  Returns a NSArray value for the specified key.
@@ -52,7 +52,7 @@
  @param keyPath A key path of the form relationship.property (with one or more relationships); for example “department.name” or “department.manager.lastName”
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid, it returns nil
  */
-- (id)objectForKeyPath:(NSString *)keyPath;
+- (id)cleverPushObjectForKeyPath:(NSString *)keyPath;
 
 /**
  Returns an object for the specified keyPath
@@ -60,7 +60,7 @@
  @param keyPath A key path of the form relationship.property, see objectForKeyPath:
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid or the result is not a NSString or can't be converted to one, it returns nil
  */
-- (NSString *)stringForKeyPath:(id)keyPath;
+- (NSString *)cleverPushStringForKeyPath:(id)keyPath;
 
 /**
  Returns an object for the specified keyPath
@@ -68,7 +68,7 @@
  @param keyPath A key path of the form relationship.property, see objectForKeyPath:
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid or the result is not a NSNumber or can't be converted to one, it returns nil
  */
-- (NSNumber *)numberForKeyPath:(id)keyPath;
+- (NSNumber *)cleverPushNumberForKeyPath:(id)keyPath;
 
 /**
  Returns an object for the specified keyPath
@@ -77,7 +77,7 @@
  @param numberFormatter The formatter to use to parse the number if the object found on the keypath is a string
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid or the result is not a NSNumber or can't be converted to one, it returns nil
  */
-- (NSNumber *)numberForKeyPath:(id)keyPath usingFormatter:(NSNumberFormatter *)numberFormatter;
+- (NSNumber *)cleverPushNumberForKeyPath:(id)keyPath usingFormatter:(NSNumberFormatter *)numberFormatter;
 
 /**
  Returns an object for the specified keyPath
@@ -85,7 +85,7 @@
  @param keyPath A key path of the form relationship.property, see objectForKeyPath:
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid or the result is not a NSArray, it returns nil
  */
-- (NSArray *)arrayForKeyPath:(id)keyPath;
+- (NSArray *)cleverPushArrayForKeyPath:(id)keyPath;
 
 /**
  Returns an object for the specified keyPath
@@ -93,7 +93,7 @@
  @param keyPath A key path of the form relationship.property, see objectForKeyPath:
  @returns The value for the derived property identified by keyPath. If the keyPath is not valid or the result is not a NSDictionary, it returns nil
  */
-- (NSDictionary *)dictionaryForKeyPath:(id)keyPath;
+- (NSDictionary *)cleverPushDictionaryForKeyPath:(id)keyPath;
 
 
 @end